fix/war: explicitly use 64-bit masks in warp sync primitives.
Details in SWDEV-490930: HIP enforces to explicitly use 64bit masks
to avoid errors when compiling for wave32 and later compiling for
wave64.

+++ cpp/src/copying/concatenate.cu
@@ -141,7 +141,7 @@ __global__ void concatenate_masks_kernel(column_device_view const* views,
 {
   auto tidx         = cudf::detail::grid_1d::global_thread_id();
   auto const stride = cudf::detail::grid_1d::grid_stride();
-  auto active_mask  = __ballot_sync(cudf::LANE_MASK_ALL, tidx < number_of_mask_bits);
+  auto active_mask  = __ballot_sync((uint64_t)cudf::LANE_MASK_ALL, tidx < number_of_mask_bits); // FIXME(HIP/AMD): WAR for SWDEV-490930
 
   size_type warp_valid_count = 0;
 
@@ -156,7 +156,7 @@ __global__ void concatenate_masks_kernel(column_device_view const* views,
       size_type const column_element_index = mask_index - output_offsets[source_view_index];
       bit_is_set = views[source_view_index].is_valid(column_element_index);
     }
-    bitmask_type const new_word = __ballot_sync(active_mask, bit_is_set);
+    bitmask_type const new_word = __ballot_sync((uint64_t) active_mask, bit_is_set); // FIXME(HIP/AMD): WAR for SWDEV-490930
 
     if (threadIdx.x % detail::warp_size == 0) {
       dest_mask[word_index(mask_index)] = new_word;
@@ -164,7 +164,7 @@ __global__ void concatenate_masks_kernel(column_device_view const* views,
     }
 
     tidx += stride;
-    active_mask = __ballot_sync(active_mask, tidx < number_of_mask_bits);
+    active_mask = __ballot_sync((uint64_t) active_mask, tidx < number_of_mask_bits); // FIXME(HIP/AMD): WAR for SWDEV-490930
   }
 
   using detail::single_lane_block_sum_reduce;
@@ -222,7 +222,7 @@ __global__ void fused_concatenate_kernel(column_device_view const* input_views,
   size_type warp_valid_count = 0;
 
   bitmask_type active_mask;
-  if (Nullable) { active_mask = __ballot_sync(cudf::LANE_MASK_ALL, output_index < output_size); }
+  if (Nullable) { active_mask = __ballot_sync((uint64_t) cudf::LANE_MASK_ALL, output_index < output_size); } // FIXME(HIP/AMD): WAR for SWDEV-490930
   while (output_index < output_size) {
     // Lookup input index by searching for output index in offsets
     auto const offset_it            = thrust::prev(thrust::upper_bound(
@@ -237,7 +237,7 @@ __global__ void fused_concatenate_kernel(column_device_view const* input_views,
 
     if (Nullable) {
       bool const bit_is_set       = input_view.is_valid(offset_index);
-      bitmask_type const new_word = __ballot_sync(active_mask, bit_is_set);
+      bitmask_type const new_word = __ballot_sync((uint64_t) active_mask, bit_is_set); // FIXME(HIP/AMD): WAR for SWDEV-490930
 
       // First thread writes bitmask word
       if (threadIdx.x % detail::warp_size == 0) {
@@ -248,7 +248,7 @@ __global__ void fused_concatenate_kernel(column_device_view const* input_views,
     }
 
     output_index += stride;
-    if (Nullable) { active_mask = __ballot_sync(active_mask, output_index < output_size); }
+    if (Nullable) { active_mask = __ballot_sync((uint64_t) active_mask, output_index < output_size); } // FIXME(HIP/AMD): WAR for SWDEV-490930
   }
 
   if (Nullable) {
