+++ cpp/src/replace/replace.cu
@@ -172,7 +172,7 @@ __global__ void replace_strings_first_pass(cudf::column_device_view input,
   auto tid              = cudf::detail::grid_1d::global_thread_id();
   auto const stride     = cudf::detail::grid_1d::grid_stride();
   cudf::bitmask_type active_mask  = cudf::LANE_MASK_ALL;
-  active_mask           = __ballot_sync(active_mask, tid < nrows);
+  active_mask           = __ballot_sync((uint64_t) active_mask, tid < nrows);
   auto const lane_id{threadIdx.x % cudf::detail::warp_size};
   uint32_t valid_sum{0};
 
@@ -197,14 +197,14 @@ __global__ void replace_strings_first_pass(cudf::column_device_view input,
       indices.data<cudf::size_type>()[idx] = -1;
     }
 
-    cudf::bitmask_type bitmask = __ballot_sync(active_mask, output_is_valid);
+    cudf::bitmask_type bitmask = __ballot_sync((uint64_t) active_mask, output_is_valid);
     if (0 == lane_id) {
       output_valid[cudf::word_index(idx)] = bitmask;
       valid_sum += cudf::__POPC(bitmask);
     }
 
     tid += stride;
-    active_mask = __ballot_sync(active_mask, tid < nrows);
+    active_mask = __ballot_sync((uint64_t) active_mask, tid < nrows);
   }
 
   // Compute total valid count for this block and add it to global count
@@ -300,7 +300,7 @@ __global__ void replace_kernel(cudf::column_device_view input,
   auto const stride = cudf::detail::grid_1d::grid_stride();
 
   cudf::bitmask_type active_mask  = cudf::LANE_MASK_ALL;
-  active_mask          = __ballot_sync(active_mask, tid < nrows);
+  active_mask          = __ballot_sync((uint64_t) active_mask, tid < nrows);
   auto const lane_id{threadIdx.x % cudf::detail::warp_size};
   uint32_t valid_sum{0};
 
@@ -323,7 +323,7 @@ __global__ void replace_kernel(cudf::column_device_view input,
 
     /* output valid counts calculations*/
     if (input_has_nulls or replacement_has_nulls) {
-      cudf::bitmask_type bitmask = __ballot_sync(active_mask, output_is_valid);
+      cudf::bitmask_type bitmask = __ballot_sync((uint64_t) active_mask, output_is_valid);
       if (0 == lane_id) {
         output.set_mask_word(cudf::word_index(idx), bitmask);
         valid_sum += cudf::__POPC(bitmask);
@@ -331,7 +331,7 @@ __global__ void replace_kernel(cudf::column_device_view input,
     }
 
     tid += stride;
-    active_mask = __ballot_sync(active_mask, tid < nrows);
+    active_mask = __ballot_sync((uint64_t) active_mask, tid < nrows);
   }
   if (input_has_nulls or replacement_has_nulls) {
     // Compute total valid count for this block and add it to global count
